requirements.txt
absl-py==2.1.0
alembic==1.13.2
aniso8601==9.0.1
asttokens==2.4.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
Deprecated==1.2.14
diffusers==0.29.2
dm-tree==0.1.8
dm_pix==0.4.3
entrypoints==0.4
etils==1.7.0
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
flax==0.8.5
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
gunicorn==22.0.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jax==0.4.30
jax-cuda12-pjrt==0.4.30
jax-cuda12-plugin==0.4.30
jaxlib==0.4.30
jaxopt==0.8.3
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
mlflow==2.14.3
mlflow-skinny==2.14.3
mlx-data==0.0.2
msgpack==1.0.8
nest-asyncio==1.6.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.4.131
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.2.1.18
nvidia-cufft-cu12==11.2.1.3
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opt-einsum==3.3.0
optax==0.2.3
orbax-checkpoint==0.5.20
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
Pygments==2.18.0
pynvml==11.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
querystring-parser==1.2.4
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scipy==1.12.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.31
sqlparse==0.5.0
stack-data==0.6.3
tensorflow-probability==0.24.0
tensorstore==0.1.63
tokenizers==0.19.1
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
